requirements.txt
torch
torchvision
hydra-core
tensorboard
rich
torchinfo
omegaconf
